refactor: GeneratedId.java 제거
