--- src/main/java/qna/domain/GeneratedId.java
@@ -1,22 +1,0 @@
-package qna.domain;
-
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.*;
-
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public abstract class GeneratedId {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Long id;
-
-    public Long getId() {
-        return id;
-    }
-
-    protected void setId(final Long id) {
-        this.id = id;
-    }
-}
