+++ src/test/java/qna/service/QnaServiceTest.java
@@ -46,8 +46,7 @@ class QnaServiceTest {
 
     @Test
     public void delete_성공() throws Exception {
-        Mockito.lenient().when(questionRepository.findByIdAndDeletedFalse(question.getId())).thenReturn(Optional.of(question));
-        Mockito.lenient().when(answerRepository.findByQuestionIdAndDeletedFalse(question.getId())).thenReturn(Arrays.asList(answer));
+        when(answerRepository.findByQuestionIdAndDeletedFalse(question.getId())).thenReturn(Arrays.asList(answer));
 
         assertThat(question.isDeleted()).isFalse();
         qnaService.deleteQuestion(UserTest.JAVAJIGI, question);
